[SG-52] style: Ui -> UI 수정
