+++ feature/profile/src/main/kotlin/com/captures2024/soongan/feature/profile/state/edit/EditProfileUIState.kt
@@ -3,7 +3,7 @@ package com.captures2024.soongan.feature.profile.state.edit
 import com.captures2024.soongan.core.analytics.utils.LogElementArgument
 import com.captures2024.soongan.core.common.base.UIState
 
-internal data class EditProfileUiState(
+internal data class EditProfileUIState(
     val isLoading: Boolean = false,
     val profileImage: String? = null,
     val nickname: String = "user1",
@@ -18,11 +18,4 @@ internal data class EditProfileUiState(
         LogElementArgument("selfIntroduction", selfIntroduction),
         LogElementArgument("isEditEnabled", isEditEnabled.toString()),
     )
-
-    override fun equals(other: Any?): Boolean {
-        return super.equals(other)
-    }
-//    operator fun compareTo() {
-//
-//    }
 }
