--- feature/profile/src/main/kotlin/com/captures2024/soongan/feature/profile/state/notification/NotificationUiState.kt
@@ -1,15 +1,0 @@
-package com.captures2024.soongan.feature.profile.state.notification
-
-import java.time.LocalDateTime
-
-internal data class NotificationUiState(
-    val isLoading: Boolean = false,
-    val notification: List<MockNotification> = emptyList(),
-)
-
-internal data class MockNotification(
-    val type: String = "",
-    val title: String = "",
-    val content: String = "",
-    val createdAt: LocalDateTime = LocalDateTime.now(),
-)
